requirements.txt
gzip
json
jsonlines
numpy
os
pandas
sys 
time
glob
torch
sentence_transformer
tensorflow
keras